Clean up/1 (#23)
* consolidate methods to base

* update project structure
