+++ Dockerfile
@@ -21,6 +21,6 @@ COPY requirements.txt  .
 RUN  pip3 install -r requirements.txt --target "${LAMBDA_TASK_ROOT}"
 
 COPY . ${LAMBDA_TASK_ROOT}
-COPY warn_classes/ ${LAMBDA_TASK_ROOT}/warn_classes/
+COPY warn/ ${LAMBDA_TASK_ROOT}/warn/
 
 CMD [ "main.handler" ]
