+++ warn/states/ca.py
@@ -2,7 +2,7 @@ import re
 from typing import List
 
 import pandas as pd
-from .base_warn import Warn
+from ..warn_base import Warn
 
 class CAWarn(Warn):
     url = 'https://edd.ca.gov/siteassets/files/jobs_and_training/warn/warn_report.xlsx'
@@ -14,9 +14,11 @@ class CAWarn(Warn):
 
     def _fetch_latest_notices(self) -> dict:
         df = pd.read_excel(self.url, sheet_name='Sheet1', dtype=str)
+        print(df.columns)
+        df['Received\nDate'] = df['Received\nDate'].astype(str)
+
         month, date, year = self.get_month_date_year(self._compare_date)
         match_date = f'{year}-{month}-{date} 00:00:00'
-        df['Received\nDate'] = df['Received\nDate'].astype(str)
         df = df[df['Received\nDate'] == match_date]
 
         layoffs = {}
@@ -27,18 +29,3 @@ class CAWarn(Warn):
                 layoffs[company_name] = 0 
             layoffs[company_name] += int(number_affected)
         return layoffs
-
-    def get_month_date_year(self, date:str):
-        date_regex = re.compile(r'(\d{1,2})/(\d{1,2})/(\d{1,4})')
-        match = date_regex.search(date)
-        if match:
-            month = match.group(1)
-            date = match.group(2)
-            year = match.group(3)
-
-            if len(month) == 1:
-                month = "0" + month
-            if len(date) == 1:
-                date = "0" + date
-
-        return month, date, year
